feat(filter): Added result in button "show"
