+++ REES46/Classes/Stories/StoryViewController/SearchWidget/Extensions/SearchFilterViewController.extension.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+extension SearchFilterViewController: SearchFilterCheckBoxViewDelegate {
+    func searchFilterCheckBoxView(
+        _ view: SearchFilterCheckBoxView,
+        didUpdateSelectedTypes selectedTypes: Set<String>,
+        header:String
+    ) {
+        selectedFilters[header] = Array(selectedTypes)
+        performSearch(with: selectedFilters)
+    }
+}
