+++ REES46/Classes/Stories/StoryViewController/SearchWidget/Protocol/CheckBoxDelegate.protocol.swift
@@ -1,0 +1,5 @@
+import UIKit
+
+protocol CheckBoxDelegate: AnyObject {
+    func checkBoxCell(didChangeState isChecked: Bool, for color: String)
+}
